move events to API bukkit.event.npc package
