+++ src/main/java/org/betonquest/betonquest/api/quest/npc/feature/NpcInteractCatcher.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.api.quest.npc.feature;
 
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.NpcInteractEvent;
+import org.betonquest.betonquest.api.bukkit.event.npc.NpcInteractEvent;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.betonquest.betonquest.id.NpcID;
